odf_odt: Test the meta.xml as well as the contents
Add an assert to the contents of meta.xml as well as content.xml. The
generated metadata was not tested for, this adds it to the testing
framework.

This requires the creation time and creation user to be faked for each
test. A helper function is added for each of these.